+++ src/modelling/multibody/gravity-compensation-residual.cpp
@@ -1,0 +1,5 @@
+#include "aligator/modelling/multibody/gravity-compensation-residual.hxx"
+
+namespace aligator {
+template struct GravityCompensationResidualTpl<context::Scalar>;
+} // namespace aligator